Chỉnh sửa va cham mario và box

+++ 05-SceneManager/PlayScene.cpp
@@ -129,7 +129,7 @@ void CPlayScene::_ParseSection_OBJECTS(string line)
 			float m = (float)atof(tokens[3].c_str());
 			float n = (float)atof(tokens[4].c_str());
 			float p = (float)atof(tokens[5].c_str());
-			float q = (float)atof(tokens[6].c_str());
+			int q = (int)atoi(tokens[6].c_str());
 			for (float i = 1; i <= m; i++)
 			{
 				if (p == 1) {
@@ -147,7 +147,7 @@ void CPlayScene::_ParseSection_OBJECTS(string line)
 		float m = (float)atof(tokens[3].c_str());
 		float n = (float)atof(tokens[4].c_str());
 		float p = (float)atof(tokens[5].c_str());
-		float q = (float)atof(tokens[6].c_str());
+		int q = (int)atoi(tokens[6].c_str());
 		for (float i = 1; i <= m; i++)
 		{
 			if (p == 1) {
@@ -165,7 +165,7 @@ void CPlayScene::_ParseSection_OBJECTS(string line)
 		float m = (float)atof(tokens[3].c_str());
 		float n = (float)atof(tokens[4].c_str());
 		float p = (float)atof(tokens[5].c_str());
-		float q = (float)atof(tokens[6].c_str());
+		int q = (int)atoi(tokens[6].c_str());
 		for (float i = 1; i <= m; i++)
 		{
 			if (p == 1) {
@@ -183,7 +183,7 @@ void CPlayScene::_ParseSection_OBJECTS(string line)
 		float m = (float)atof(tokens[3].c_str());
 		float n = (float)atof(tokens[4].c_str());
 		float p = (float)atof(tokens[5].c_str());
-		float q = (float)atof(tokens[6].c_str());
+		int q = (int)atoi(tokens[6].c_str());
 		for (float i = 1; i <= m; i++)
 		{
 			if (p == 1) {
@@ -201,7 +201,7 @@ void CPlayScene::_ParseSection_OBJECTS(string line)
 		float m = (float)atof(tokens[3].c_str());
 		float n = (float)atof(tokens[4].c_str());
 		float p = (float)atof(tokens[5].c_str());
-		float q = (float)atof(tokens[6].c_str());
+		int q = (int)atoi(tokens[6].c_str());
 		for (float i = 1; i <= m; i++)
 		{
 			if (p == 1) {
@@ -219,7 +219,7 @@ void CPlayScene::_ParseSection_OBJECTS(string line)
 		float m = (float)atof(tokens[3].c_str());
 		float n = (float)atof(tokens[4].c_str());
 		float p = (float)atof(tokens[5].c_str());
-		float q = (float)atof(tokens[6].c_str());
+		int q = (int)atoi(tokens[6].c_str());
 		for (float i = 1; i <= m; i++)
 		{
 			if (p == 1) {
@@ -237,7 +237,7 @@ void CPlayScene::_ParseSection_OBJECTS(string line)
 		float m = (float)atof(tokens[3].c_str());
 		float n = (float)atof(tokens[4].c_str());
 		float p = (float)atof(tokens[5].c_str());
-		float q = (float)atof(tokens[6].c_str());
+		int q = (int)atoi(tokens[6].c_str());
 		for (float i = 1; i <= m; i++)
 		{
 			if (p == 1) {
